+++ src/activities/reader/WordRef.h
@@ -2,6 +2,7 @@
 
 #include <EpdFontFamily.h>
 
+#include <cstdint>
 #include <string>
 
 struct WordRef {
@@ -10,8 +11,10 @@ struct WordRef {
   int w = 0;
   int h = 0;
   int pageIdx = 0;
+  uint16_t pageWordIndex = 0;
   std::string text;
   EpdFontFamily::Style style = EpdFontFamily::REGULAR;
   bool paragraphStart = false;
   bool endsWithInsertedHyphen = false;
+  bool lineIsRtl = false;
 };